MarketInsightAgency/requirements.txt
agency-swarm
pandas
numpy
matplotlib
jinja2
python-dotenv
gradio
pytrends
requests
statista
psycopg2-binary
google-search-results 